working wasm tracing
