+++ crates/hdk3/Cargo.toml
@@ -24,10 +24,11 @@ serde = "1.0.104"
 serde_bytes = "0.11"
 thiserror = "1.0.22"
 tracing = "0.1"
+tracing-core = "0.1"
 
 [dependencies.tracing-subscriber]
 version = "0.2"
-features = ["fmt", "ansi"]
+features = []
 
 [dev-dependencies]
 fixt = { path = "../fixt" }
