+++ crates/hdk3/src/host_fn.rs
@@ -3,7 +3,6 @@ pub mod call;
 pub mod call_remote;
 pub mod create;
 pub mod create_link;
-pub mod trace;
 pub mod delete;
 pub mod delete_link;
 pub mod emit_signal;
@@ -21,6 +20,7 @@ pub mod schedule;
 pub mod show_env;
 pub mod sign;
 pub mod sys_time;
+pub mod trace;
 pub mod unreachable;
 pub mod update;
 pub mod verify_signature;
