+++ crates/hdk3/src/prelude.rs
@@ -3,7 +3,6 @@ pub use crate::capability::create_cap_grant::create_cap_grant;
 pub use crate::capability::delete_cap_grant::delete_cap_grant;
 pub use crate::capability::generate_cap_secret::generate_cap_secret;
 pub use crate::capability::update_cap_grant::update_cap_grant;
-pub use crate::debug;
 pub use crate::entry::create_entry::create_entry;
 pub use crate::entry::delete_entry::delete_entry;
 pub use crate::entry::hash_entry::hash_entry;
@@ -61,6 +60,7 @@ pub use holochain_zome_types::prelude::*;
 pub use std::collections::HashSet;
 pub use std::convert::TryFrom;
 pub use tracing;
+pub use tracing::{debug, error, info, instrument, trace, warn};
 pub use tracing_subscriber;
 
 // This needs to be called at least once _somewhere_ and is idempotent.
