+++ crates/holochain_zome_types/src/prelude.rs
@@ -9,8 +9,6 @@ pub use crate::capability::*;
 pub use crate::cell::*;
 pub use crate::cell::*;
 pub use crate::crdt::*;
-pub use crate::trace::*;
-pub use crate::trace_msg;
 pub use crate::element::*;
 pub use crate::entry::*;
 pub use crate::entry::*;
@@ -30,6 +28,7 @@ pub use crate::request::*;
 pub use crate::signal::*;
 pub use crate::signature::*;
 pub use crate::timestamp::*;
+pub use crate::trace::*;
 pub use crate::validate::*;
 pub use crate::validate_link::*;
 pub use crate::warrant::*;
